+++ lib/rubocop.rb
@@ -202,6 +202,7 @@ require 'rubocop/cop/style/while_until_do'
 require 'rubocop/cop/style/while_until_modifier'
 require 'rubocop/cop/style/word_array'
 
+require 'rubocop/cop/rails/action_filter'
 require 'rubocop/cop/rails/default_scope'
 require 'rubocop/cop/rails/has_and_belongs_to_many'
 require 'rubocop/cop/rails/output'